Enhancement to support the use of a persistent etcd serviceinfo key. The etcd serviceinfo key was composed of a sessionId + '.' + serviceinfoid.   Both the sessionId and the serviceinfoid were random UUIDs with UUID string syntax.   The sessionid can now be set via a system property:
-Decf.discovery.etcd.sessionid=<sessionid>

And the serviceinfoid will be set to the String value of the
endpoint.id, which for remote services is the OSGi standard service
property that uniquely identifies the remote service endpoint
description.   If not explicitly set, the endpoint.id is set by ECF RSA
to a random UUID.   If explicitly set for a given remote service, it
should have UUID syntax.